requirements.txt
pytest==7.4.3
requests==2.31.0
pydantic==1.10.13  # Более старая, но стабильная версия
python-dotenv==1.0.0
faker==20.1.0
allure-pytest==2.13.2